Refactor: 클래스가 하나의 역할을 하도록 MVC 패턴 적용
